[jest-expo] Create a multi-project config that runs iOS, Android, and Web (#5006)
* Create a multi-project config that runs iOS, Android, and Web

* Created `jest-expo/universal`

* Removed React Native from `web` env

* Added universal to module scripts

* Updated format

* Update babel.config.js

* Update packages/jest-expo/babel.config.js


* Renamed test files

* Update jest-preset.js

* Update jest-preset.js
