+++ packages/expo-module-scripts/universal/jest-preset.js
@@ -1,0 +1,2 @@
+const createJestPreset = require('../createJestPreset');
+module.exports = createJestPreset(require('jest-expo/universal/jest-preset'));
