+++ packages/jest-expo/babel.config.js
@@ -1,0 +1,11 @@
+module.exports = function(api) {
+  api.cache(true);
+  return {
+    // Only use this when running tests
+    env: {
+      test: {
+        presets: ['babel-preset-expo'],
+      },
+    },
+  };
+};
