+++ packages/jest-expo/tests/default-extension.android.js
@@ -1,0 +1,1 @@
+export const OS = ['android'];
