+++ packages/jest-expo/tests/default-extension.ios.js
@@ -1,0 +1,1 @@
+export const OS = ['ios'];
