+++ packages/jest-expo/tests/default-extension.web.js
@@ -1,0 +1,1 @@
+export const OS = ['web'];
