+++ packages/jest-expo/tests/fallback-extension.js
@@ -1,0 +1,1 @@
+export const OS = ['web'];
