+++ packages/jest-expo/tests/fallback-extension.native.js
@@ -1,0 +1,1 @@
+export const OS = ['ios', 'android'];
