+++ packages/jest-expo/tests/fallback-extension.node.js
@@ -1,0 +1,1 @@
+export const OS = ['web'];
